refactor: 버튼 폴더 정리
