+++ client/src/components/phaser/map/map.tsx
@@ -6,10 +6,10 @@ import Phaser from "phaser";
 import { MeetsInPhaserScene } from "./phaserScene";
 import { phaserConfig } from "./phaserConfig";
 import { io } from "socket.io-client";
+import MapZoomButtons from "./mapZoom/mapZoomButtons";
 import { useGetUserInfo } from "@/apis/service/user.service";
 import { useAtomValue } from "jotai";
 import { isChatFocusedAtom, zoomLevelAtom } from "@/jotai/atom";
-import MapZoomButtons from "@/components/button/mapZoom/mapZoomButtons";
 import style from "./map.module.scss";
 
 const Map = () => {
