+++ client/src/components/phaser/map/mapZoom/mapZoomButtons.module.scss
@@ -1,5 +1,5 @@
-@import "../../../styles/mixin";
-@import "../../../styles/variables";
+@import "../../../../styles/mixin";
+@import "../../../../styles/variables";
 
 .container {
     @include common_ui;
