+++ client/src/components/room/roomContent.tsx
@@ -5,15 +5,15 @@ import { useParams } from "next/navigation";
 import { useAtomValue } from "jotai";
 import { screenShareStateAtom } from "@/jotai/atom";
 import { ScreenShareState } from "@/types/peer.type";
-import { useGetRoomData } from "@/apis/service/room.service";
-import useChatSocket from "@/app/room/[roomId]/hooks/useChatSocket";
 import { useScreenShare } from "@/app/room/[roomId]/hooks/useScreenShare";
+import { useGetRoomData } from "@/apis/service/room.service";
 import Chat from "@/components/chat/chat";
 import Menu from "@/components/menu/menu";
+import RoomGradientBackground from "@/components/background/room/roomGradientBackground";
 import ScreenWindow from "@/components/screen/window/screenWindow";
 import Skeleton from "@/components/common/skeleton/skeleton";
-import ViewSwitchButton from "@/components/buttons/viewSwitch/viewSwitchButton";
-import RoomGradientBackground from "@/components/background/room/roomGradientBackground";
+import ViewSwitchButton from "../screen/viewSwitch/viewSwitchButton";
+import useChatSocket from "@/app/room/[roomId]/hooks/useChatSocket";
 import { notFound } from "next/navigation";
 import style from "./roomContent.module.scss";
